Updated Editor version to be compatible with newer Unity VR tools
In order to get the new Unity XR plugin management system to work in the project, I updated the editor version to 2020.3.26f1. Nothing seems broken as a result, and the scene should now have functioning teleportation in VR. It should also be more adaptable for any VR headset now because of the update. It should also be testable without a VR headset plugged in.